requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Limiter==1.4
Flask-RESTful==0.3.8
Flask-PyMongo==2.3.0
dnspython==2.0.0
PyJWT==1.7.1
